requirements.txt
Flask
langchain
langchain-community
langchain-huggingface
langchain-openai
faiss-cpu
sentence-transformers
python-dotenv
langchain-groq
langchain-google-genai
